frob: move safety/auth into vat
Consolidate the basic CDP safety and ownership conditions into the Vat.

`frob` is the basic CDP operation and governs the creation of Dai on a
fundamental level, according to the basic principle of excess collateral
backing, and is the core domain of the system.

The safety and ownership mechanisms for CDPs were split out into a
separate contract, the Pit, as part of an exploratory design process,
but are now reconsolidated.