vat: use address rather than bytes32 for data
The intent behind using bytes32 was to give addresses maximum
flexibility in administering CDPs, with the specific use case of a CDP
manager having control over a full 12 bytes of CDPs on behalf of users.

This model was used prior to the addition of CDP permissioning and
`fork`, which were required to allow for easier interaction between
users and different management contracts.

With the addition of these features bytes32 becomes less justifiable as
a choice and the restriction of a single address to a single urn per ilk
becomes attractive.